neutron_reproduction_12x12/requirements.txt
numpy
numpy-quaternion
torch
matplotlib
tqdm
scipy
setuptools
